# crates/nexis-gateway/src/router/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Wisdoverse Nexus Gateway API",
    "version": "0.1.0",
    "description": "Control plane API for room management, messaging, semantic search, privacy self-service, and collaboration."
  },
  "paths": {
...
    "/v1/rooms": {
      "get": {
        "summary": "List rooms",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Authorized rooms only",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomList"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, expired, or tenant-scoped credentials"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "403": {
            "description": "Member is not authorized for this room operation"
          },
          "404": {
            "description": "Room does not exist"
          },
          "503": {
            "description": "Unsupported multi-tenant mode or retry capacity exhausted"
          }
        }
      },
      "post": {
        "summary": "Create room",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Room created with JWT subject as owner",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomCreated"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "description": "Missing, invalid, expired, or tenant-scoped credentials"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "503": {
            "description": "Unsupported multi-tenant mode or retry capacity exhausted"
          },
          "403": {
            "description": "Member is not authorized for this room operation"
          },
          "404": {
            "description": "Room does not exist"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomCreate"
              }
            }
          }
        }
      }
    },
    "/v1/rooms/{id}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/RoomId"
        }
      ],
      "get": {
        "summary": "Get room details",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Authorized room and history",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomDetails"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "description": "Missing, invalid, expired, or tenant-scoped credentials"
          },
          "404": {
            "description": "Room does not exist"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "503": {
            "description": "Unsupported multi-tenant mode or retry capacity exhausted"
          },
          "403": {
            "description": "Member is not authorized for this room operation"
          }
        }
      },
      "delete": {
        "summary": "Delete room",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Room deleted"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "description": "Missing, invalid, expired, or tenant-scoped credentials"
          },
          "404": {
            "description": "Room does not exist"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "503": {
            "description": "Unsupported multi-tenant mode or retry capacity exhausted"
          },
          "403": {
            "description": "Member is not authorized for this room operation"
          }
        }
      }
    },
    "/v1/rooms/{id}/invite": {
      "parameters": [
        {
          "$ref": "#/components/parameters/RoomId"
        }
      ],
      "post": {
        "summary": "Invite member to room",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Member invited"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "description": "Missing, invalid, expired, or tenant-scoped credentials"
          },
          "404": {
            "description": "Room does not exist"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "503": {
            "description": "Unsupported multi-tenant mode or retry capacity exhausted"
          },
          "403": {
            "description": "Member is not authorized for this room operation"
          }
        }
      }
    },
    "/v1/messages": {
      "post": {
        "summary": "Send message",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Committed message; same key and payload return the same identity without another broadcast",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "400": {
            "description": "Blank/oversize text, invalid retry key, or reply outside this room"
          },
          "401": {
            "description": "Missing, invalid, expired, or tenant-scoped credentials"
          },
          "404": {
            "description": "Room does not exist"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "503": {
            "description": "Unsupported multi-tenant mode or retry capacity exhausted"
          },
          "403": {
            "description": "Member is not authorized for this room operation"
          },
          "409": {
            "description": "Retry key reused with different text or reply reference"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageCreate"
              }
            }
          }
        }
      }
    },
    "/v1/search": {
      "get": {
        "summary": "Search messages with query parameters",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Search response"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      },
      "post": {
        "summary": "Search messages with JSON body",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Search response"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/v1/members/me/export": {
      "get": {
        "summary": "Export authenticated member data",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Member data export"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/v1/members/me": {
      "delete": {
        "summary": "Delete authenticated member data",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "202": {
            "description": "Deletion accepted"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/v1/collaboration/meetings/rooms": {
      "post": {
        "summary": "Create meeting room",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ApiVersion"
          }
        ],
        "responses": {
          "201": {
            "description": "Meeting room created"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/v1/collaboration/meetings/rooms/{room_id}/join": {
      "parameters": [
        {
          "$ref": "#/components/parameters/CollaborationRoomId"
        }
      ],
      "post": {
        "summary": "Join meeting room",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ApiVersion"
          }
        ],
        "responses": {
          "200": {
            "description": "Joined meeting room"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/v1/collaboration/meetings/rooms/{room_id}/leave": {
      "parameters": [
        {
          "$ref": "#/components/parameters/CollaborationRoomId"
        }
      ],
      "post": {
        "summary": "Leave meeting room",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ApiVersion"
          }
        ],
        "responses": {
          "200": {
            "description": "Left meeting room"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "parameters": {
...
      "RoomId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Validation or request contract error"
      },
...
      "RateLimited": {
        "description": "Rate limit exceeded"
      },
...
    },
    "schemas": {
...
      "RoomCreated": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
...
      "RoomList": {
        "type": "object",
        "properties": {
          "rooms": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoomSummary"
            }
          },
          "total": {
            "type": "integer",
            "minimum": 0
          }
        },
        "required": [
          "rooms",
          "total"
        ]
      },
...
      "RoomCreate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "topic": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "name"
        ]
      }
    }
  }
}